Added create employee functionality

+++ database/migrations/2023_05_26_141357_create_employees_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->dateTime('dob');
             $table->string('gender');
             $table->dateTime('join_date');
-            $table->float('salary');
+            $table->decimal('salary', 10, 2);
             $table->timestamps();
         });
     }
